feat(apps-script): test reorganization
