+++ packages/apps-script/src/_test/_test.js
@@ -1,0 +1,8 @@
+function test_All() {
+    test_Utils();
+    test_Sheet();
+    test_Entity();
+    test_I18n();
+    test_Api();
+    test_App();
+}
